Revert "courses new UI"
