+++ package.json
@@ -8,27 +8,13 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@expo/vector-icons": "^12.0.0",
-    "@react-native-community/masked-view": "^0.1.7",
-    "@react-navigation/bottom-tabs": "^5.4.5",
-    "@react-navigation/drawer": "^5.7.5",
-    "@react-navigation/material-bottom-tabs": "^5.2.5",
-    "@react-navigation/material-top-tabs": "^5.2.5",
-    "@react-navigation/native": "^5.4.0",
-    "@react-navigation/stack": "^5.3.7",
+    "@react-navigation/native": "^6.0.6",
+    "@react-navigation/native-stack": "^6.2.5",
     "expo": "~43.0.0",
     "expo-status-bar": "~1.1.0",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "0.64.2",
-    "react-native-gesture-handler": "~1.10.2",
-    "react-native-paper": "^3.10.1",
-    "react-native-reanimated": "~2.2.0",
-    "react-native-safe-area-context": "3.3.2",
-    "react-native-screens": "~3.8.0",
-    "react-native-tab-view": "^2.14.0",
-    "react-native-vector-icons": "^9.0.0",
-    "react-native-web": "0.17.1"
     "react-native-heroicons": "^2.0.2",
     "react-native-web": "0.17.1",
     "tailwind-react-native-classnames": "^1.5.1"
