--- pages/DetailsScreen.js
@@ -1,43 +1,0 @@
-import * as React from 'react';
-import { View, Text, SafeAreaView } from 'react-native';
-
-const DetailsScreen = () => {
-  return (
-    <SafeAreaView style={{ flex: 1 }}>
-      <View style={{ flex: 1 , padding: 16}}>
-        <View
-          style={{
-            flex: 1,
-            alignItems: 'center',
-            justifyContent: 'center',
-          }}>
-          <Text
-            style={{
-              fontSize: 25,
-              textAlign: 'center',
-              marginBottom: 16
-            }}>
-            You are on Details Screen
-          </Text>
-        </View>
-        <Text
-          style={{
-            fontSize: 18,
-            textAlign: 'center',
-            color: 'grey'
-          }}>
-           The mentor app!!!!
-        </Text>
-        <Text
-          style={{
-            fontSize: 16,
-            textAlign: 'center',
-            color: 'grey'
-          }}>
-          Lets get coding team
-        </Text>
-      </View>
-    </SafeAreaView>
-  );
-}
-export default DetailsScreen;
