--- pages/Profile.js
@@ -1,27 +1,0 @@
-import { StatusBar } from 'expo-status-bar';
-import React from 'react';
-import { StyleSheet, Image, Text, View, SafeAreaView } from 'react-native';
-
-
-class Profile extends React.Component{
-  render() {
-    return(
-<View style={styles.container}>
-<Text>Go to Profile!</Text>
-</View>
-    );
-  }
-}
-export default Profile;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
-
-
-
